Substantial change in reading logic. See extended commit message.
It turns out, OpenFOAM samples quantities of the same dimension into the
same file. For example, if `T` and `p_rgh` are requested, then files of
the type `vmasts_h50_T_p_rgh` are created. The order is specified by the
requested order in controlDict. Before, the code would look for
individual files for each variable in each time-step directory. The new
behavior is that now it tries to look for the single-variable file and
if it found is follows like before; but if it is not, it looks for that
variable in different files within the same time-step directory. The
order of the printed variables are given by the order in the name of the
file and that is considered too.