Reader now works for for generic data varying in any dimension
Reader is now general so that it is not only useful just for slices.
Any VTK containing data varying in all 3 dimensions can now be read. To
resort to the slice reading, the user has to specify the `squash` flag.
Previous behavior was to try to guess in which plane the data lie even
if `squasn` was not provided. Now, without `squash`, it reads the full
3-D field.